added a test for invalid_user getting other user's portfolio
